requirements.txt
pandas
requests
SQLAlchemy
python-dotenv